account_invoice_refund_link: Don't choke with standard fields
Since v11, Odoo counts with a field called `refund_invoice_ids` of type
one2many, which conflicts with the m2m field defined in this module.
Moreover, these fields are not needed anymore.

There is only 2 OCA modules using this module right now, and they
don't need significant changes for this adaptation, but I will do it
when this get merged.